move def to server
