+++ server/encoding/yaml/definition/testdata/test_http.yaml
