+++ server/encoding/yaml/definition/testdata/transaction.yaml
